% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TENET.ExperimentHub-package.R
\docType{package}
\name{TENET.ExperimentHub}
\alias{TENET.ExperimentHub-package}
\alias{TENET.ExperimentHub}
\title{TENET.ExperimentHub (Experiment data for the TENET package)}
\description{
The TENET.ExperimentHub package contains datasets for use in
the TENET package's vignette and function examples. These include a variety
of different objects to illustrate different datasets used in TENET
functions. Where applicable, all datasets are aligned to the hg38 human
genome.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rhielab/TENET.ExperimentHub}
  \item Report bugs at \url{https://github.com/rhielab/TENET.ExperimentHub/issues}
}

}
\author{
\strong{Maintainer}: Rhie Lab at the University of Southern California \email{rhielab@gmail.com}

Authors:
\itemize{
  \item Daniel Mullen (\href{https://orcid.org/0000-0002-7639-0549}{ORCID})
  \item Zexun Wu (\href{https://orcid.org/0000-0003-2566-1326}{ORCID})
  \item Ethan Nelson-Moore (\href{https://orcid.org/0009-0001-6903-9232}{ORCID})
  \item Suhn Rhie (\href{https://orcid.org/0000-0002-5522-5296}{ORCID})
}

}
\keyword{internal}
