% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{exampleTENETStep1MakeExternalDatasetsGRanges}
\alias{exampleTENETStep1MakeExternalDatasetsGRanges}
\title{Example TENET \code{step1MakeExternalDatasets} GRanges object}
\usage{
exampleTENETStep1MakeExternalDatasetsGRanges(metadata = FALSE)
}
\arguments{
\item{metadata}{If TRUE, retrieve the ExperimentHub metadata instead of the
object itself. Defaults to FALSE.}
}
\value{
A GRanges object with 1,971,031 ranges and no metadata.
}
\description{
A GenomicRanges dataset representing putative enhancer regions
relevant to BRCA, created using the \code{step1MakeExternalDatasets} function in
the TENET package with the \code{consensusEnhancer}, \code{consensusNDR},
\code{publicEnhancer}, \code{publicNDR}, and \code{ENCODEdELS} arguments all set to TRUE,
and the \code{cancerType} argument set to "BRCA". The data are aligned to the
human hg38 genome. This dataset is included to demonstrate TENET's
\code{step2GetDifferentiallyMethylatedSites} function.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
exampleTENETStep1MakeExternalDatasetsGRanges()
\dontshow{\}) # examplesIf}
}
